Separate the parser and generator namespaces
Presence of a parser doesn't necessarily mean that the generator is
available (and vice versa).

Therefore the output of `CBF::formats` was misleading: it would say that
for instance `:heat` is available even though we only provide the Heat
generator at the moment.

This splits the namespace to `CBF::Parsers` and `CBF::Generators` along
with appropriate API changes.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>