Heat: generate user data script
This will take all the services' executables and make sure they are
marked as such in the generated Heat template.

We save the executables as regular files using the
Metadata/AWS::CloudFormation::Init/config/files directive and then
invoke the executables via a custom script put into UserData.

That way, we can update the custom stript to pass the parameters as
environment variables that the scripts will expect if coming from
Audrey.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>