Add test for requiring keynames in Aeolus parser
The Aeolus deployable format doesn't have a way to specify a key name of
the instances it describes.

Heat does, however, and it would make sense to be able to generate a
Heat template with the instance keynames as parameters.

This tests for a `require_instance_keys` option in the Aeolus parser. If
it's set, the parsed resources should have parameters set for the keys.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>